ai-workflow-automation/requirements.txt
pyyaml>=6.0
litellm>=1.30.0
aiohttp>=3.9.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
